accessToken wird beim Aufrufen des admin-panels überprüft
