+++ Prototyp/backend/server.js
@@ -342,25 +342,25 @@ import jwt from 'jsonwebtoken';
 app.use(express.json());
 
 //Sollten am besten in einer Datenbank gespeichert werden.
-const posts = [
-    {Firmenname: 'Apple', maxFileInKB: 1000, maxWidthInPX: 1000},
-    {Firmenname: 'Google', maxFileInKB: 2000, maxWidthInPX: 2000},
-    {Firmenname: 'Microsoft', maxFileInKB: 3000, maxWidthInPX: 3000}
+const customers = [
+    {username: 'Edgar', title: 1000},
+    {username: 'Antonio', title: 2000}
 ]
 
-app.get('/posts', authenticateToken, (req, res) =>{
-    // Hier werden Posts angezeigt, die alle Admins angelegt haben. Wenn nur bestimmte angezeigt werden sollenn, dann filtern.
-    res.json(posts)
+app.get('/customers', authenticateToken, (req, res) =>{
+    // Hier werden customers angezeigt, die alle Admins angelegt haben. Wenn nur bestimmte angezeigt werden sollenn, dann filtern.
+    console.log(res.json(customers))
+
 })
 
-app.post('/login', (req, res) =>{
-    //Authenticate User
-    const username = req.body.username
-    const user = {name: username}
+// app.post('/login', (req, res) =>{
+//     //Authenticate User
+//     const username = req.body.username
+//     const user = {name: username}
 
-    const accessToken = jwt.sign(user, process.env.ACCESS_TOKEN_SECRET)
-    res.json({accessToken: accessToken})
-})
+//     const accessToken = jwt.sign(user, process.env.ACCESS_TOKEN_SECRET)
+//     res.json({accessToken: accessToken})
+// })
 
 function authenticateToken(req, res, next){
     const authHeader = req.headers['authorization']
@@ -378,6 +378,67 @@ function authenticateToken(req, res, next){
     })
 }
 
+app.get('/verify-token', authenticateToken, (req, res) => {
+    // If we get here, the token is valid (authenticateToken middleware passed)
+    res.status(200).json({ valid: true });
+});
+
+
+//Authentifizierung
+
+// import dotenv from 'dotenv';
+// import express from 'express';
+// import jwt from 'jsonwebtoken';
+// import cors from 'cors';  // Füge cors import hinzu
+
+let refreshTokens = [];
+
+// Aktiviere CORS
+app.use(cors());
+app.use(express.json());
+
+app.post('/token', (req, res) =>{
+    const refreshToken = req.body.token
+    if(refreshToken == null) return res.sendStatus(401)
+    if(!refreshTokens.includes(refreshToken)) return res.sendStatus(403)
+    jwt.verify(refreshToken, process.env.REFRESH_TOKEN_SECRET, (err, user) =>{
+        if(err) return res.sendStatus(403)
+        const accessToken = generateAccessToken({name: user.name})
+        res.json({accessToken: accessToken})
+    })
+})
+
+app.delete('/logout', (req, res) =>{
+    refreshTokens = refreshTokens.filter(token => token !== req.body.token)
+    res.sendStatus(204)
+})
+
+app.post('/login', (req, res) =>{
+    //Authenticate User
+    const username = req.body.username
+    const user = {name: username}
+    console.log('-------------------------'+user)
+
+    const accessToken = generateAccessToken(user)
+    const refreshToken = jwt.sign(user, process.env.REFRESH_TOKEN_SECRET)
+    refreshTokens.push(refreshToken)
+    res.json({accessToken: accessToken, refreshToken: refreshToken})
+})
+
+function generateAccessToken(user){
+    return jwt.sign(user, process.env.ACCESS_TOKEN_SECRET, {expiresIn: '3s'})
+}
+
+
+
+
+
+
+
+
+
+
+
 app.listen(PORT, () =>
-    console.log(`Server listening on port ${PORT}`),
+    console.log(`Server listening on port ${PORT}`)
 );
