+++ Prototyp/frontend/admin-panel.html
@@ -30,7 +30,7 @@
             <a href="#"> <h2 class="hover-light-blue">Account</h2> </a>
           </li>
           <li>
-            <a href="#" class = "logoutButton"> <h2 class="hover-light-blue">Logout</h2> </a>
+            <a href="index.html" class = "logoutButton"> <h2 class="hover-light-blue">Logout</h2> </a>
         </ul>
       </nav>
     </header>
@@ -92,5 +92,65 @@
       </div>
     </main>
     <footer></footer>
+    <script>
+      //Abfangen des Requests um auf diese Seite zu gelangen
+      document.addEventListener('DOMContentLoaded', () => {
+        const accessToken = localStorage.getItem('accessToken');
+        const refreshToken = localStorage.getItem('refreshToken');
+        
+        // If there's no token at all, redirect to login immediately
+        if (!accessToken) {
+            window.location.href = '/login.html';
+            return;
+        }
+        
+        // Verify token with server
+        fetch('http://localhost:5000/verify-token', {
+            method: 'GET',
+            headers: {
+                'Authorization': `Bearer ${accessToken}`
+            }
+        })
+        .then(response => {
+            if (response.status !== 200) {
+                // If token is invalid, try refreshing
+                if (refreshToken) {
+                    return fetch('http://localhost:5000/token', {
+                        method: 'POST',
+                        headers: {
+                            'Content-Type': 'application/json'
+                        },
+                        body: JSON.stringify({ token: refreshToken })
+                    }).then(refreshResponse => {
+                        if (refreshResponse.ok) {
+                            return refreshResponse.json();
+                        } else {
+                            throw new Error('Failed to refresh token');
+                        }
+                    }).then(data => {
+                        // Store new access token
+                        localStorage.setItem('accessToken', data.accessToken);
+                        // Reload page to retry with new token
+                        window.location.reload();
+                    }).catch(error => {
+                        // If refresh fails, redirect to login
+                        console.error('Token refresh failed:', error);
+                        localStorage.removeItem('accessToken');
+                        localStorage.removeItem('refreshToken');
+                        window.location.href = '/login.html';
+                    });
+                } else {
+                    // No refresh token available, redirect to login
+                    localStorage.removeItem('accessToken');
+                    window.location.href = '/login.html';
+                }
+            }
+        })
+        .catch(error => {
+            console.error('Error verifying token:', error);
+            window.location.href = '/login.html';
+        });
+    });
+    </script>
   </body>
 </html>
