+++ Prototyp/frontend/index.html
@@ -31,7 +31,7 @@
         <ul>
           <li><a href="#"> <h2 class="hover-light-blue">Credits anfragen</h2> </a></li>
           <li><a href="#"> <h2 class="hover-light-blue">Kontakt</h2></a></li>
-          <li><a href="http://127.0.0.1:5500/Prototyp/frontend/login.html"> <h2 class="hover-light-blue">Admin</h2></a></li>
+          <li><a href="login.html"> <h2 class="hover-light-blue">Admin</h2></a></li>
         </ul>
       </nav>
     </header>
